fastapi/requirements.txt
annotated-types==0.6.0
anyio==4.2.0
bottle==0.12.25
cffi==1.16.0
click==8.1.7
clr-loader==0.2.6
colorama==0.4.6
exceptiongroup==1.2.0
fastapi==0.108.0
h11==0.14.0
httptools==0.6.1
idna==3.6
Jinja2==3.1.2
MarkupSafe==2.1.3
proxy-tools==0.1.0
pycparser==2.21
pydantic==2.5.3
pydantic-core==2.14.6
python-dotenv==1.0.0
pythonnet==3.0.3
pywebview==4.4.1
PyYAML==6.0.1
sniffio==1.3.0
starlette==0.32.0.post1
typing-extensions==4.9.0
uvicorn==0.25.0
watchfiles==0.21.0
websockets==12.0
